Change  Start location Marker to Map Pin Shape
The previous 64x64 start location circle was placed counterintuitively. It was placed so its imaginary surrounding square had its left corner at the target coordinate. This was changed so that the bottom corner instead defines the coordinate (like with islands), while the shape was changed to a map pin to more accurately denote the placement location.