Allow for Off-Center Playable Area
Found when the Playable area is an uneven number of Tiles or when the Playable area is shifted on purpose (eg. DLC12 Campaign Map).

When manually adjusting the Margin, it will only allow centered values when in centered mode, thus using Steps of 8 for the Margin size, to keep the playable area properly centered. When in custom mode, the step size for the average margin turns to 4, and the Offsets get adjusted accordingly if needed.

Loading maps that have an Offset by default will open the custom size control by default. Otherwise, the centered only control will be shown.