Mod Export for different Sessions
Can now Export a mod for the New World. During mod export you can only select map templates for override that match the session (so in the OW, you can still choose to override Corners or Island Arc, but in the NW, you can only override the general NW).

For Extended Maps (only New World currently), separate files and ModOps for "normal" and "enlarged" are generated (the game crashes otherwise).