Fix UnitTests when Game Files are not available
Some UnitTests (the RoundTrip Tests) require the game files to be present on the system. This is not the case for some environments (eg. CI builds).
These tests are now skipped when the game files are not found.

(This also causes changes in the details of how the Settings class is initialized, though the previous behaviour should be unchanged.)