Move PlayableArea scaling to separate control
The scaling of the PlayableArea with sliders is very unintuitive. Instead of that, now there is a "Edit Playable Area" Mode which gets enabled in the Session Properties. In that mode, the Playable Area can be dragged and resized (by draggable bars) on the MapView.
There is also an option to display the margins on each side of the playable area for more precise positioning.