FixedIslands show their fertilities when clicked
FixedIslands only showed the fertilities in the FixedIslandProperties when first setting or when changing the Fertilities. When clicking an island with already selected fertilities, it would show that fertility is not randomized but not show the icons. This is now fixed by loading the fertility assets on creation of the viewmodel once.