Avoid crash when loading template early
The app used to crash when a template file was loaded before the IslandRepository was finished loading (aka the gamedata was parsed fully).

This change defers the loading of all asset data of FixedIslands until the IslandRepository finishes loading. Until then, empty tiles are shown, which are labelled with the island's filename, if the island has no label set explicitly. This placeholder filename label is removed after the actual assets are loaded, because then an image will be shown.