First undo experiments
Implemented a undo-stack to experiment with undo feature. Currently, the entire MapTemplate gets serialized and pushed on the stack. Practically functional, but can get slow for complex maps. Some different solution should be looked into for the long run. But that will probably take some extensive refactoring to enable some kind of diffing.