Implemented proper UndoRedoStack
Removed the hard copy of the entire template. This was just an ugly solution for the undo/redo feature.
Now there is a proper Undo/Redo Stack that can take new entries of the IStackEntry interface. Each kind of action can have its own implementation and push it to the stack. Undo and redo is then handled by the UndoRedoStack Class.