Added undo/redo for map creation and deletion
The Stack entries are created in MapView.xaml.cs which isn't really that elegant. But that file is already quite complex and handles lots of editor interactions. So touching it not that mutch is probably a good idea for now. Ultimately, the acutal edit actions should be lifted out to have consistent methods of editing the map when the user does it by different means (like maybe toolbar or context menu) in the future.

There are still some places that seem to not be NULL-safe. Have to clean that up at some point...