Added undo/redo for session region
Also added some preparations for undo/redo of map size configuration. The way the configuration of the playable area works right now is a bit complicated and does not work well with undo/redo. This will require some more extensive reworks on that feature. But I wanted to make this more accessible anyway.