Created a custom control for Toolbar
The toolbar now has its own control with an associated service. This allows fore more flexible integration of the toolbar buttons in the editor by just providing an event to listen to. Also button states can be set globally depending on state.

Some parts are still a bit hacky, but it should be a good basis.