Made the start window game-aware
As a first step, the start window shows the detected game, or "Unsupported Game" if the game is not supported by the editor. Validating the game path determines what game is selected and assigns the Game-property in DataManager.cs. From there on the editor can use the correct methods to load each games data.