All static assets are now contained withing game classes
The Editor is now able to load 1800 or 117. For 1800, it works as normal. For 117, it can open maps from a7tinfo files.

Still quite a lot hardcoded stuff, but this now references the game classes. This needs to be worked on, especially for Mod Export.

Complicated things to look at: Island Sizes, Fertilities, loading of templates with game-dependant data models...